+++ integration-test/Analysis/LicenseScannerSpec.hs
@@ -26,6 +26,7 @@ import Srclib.Types (
   emptyLicenseUnit,
  )
 import Test.Hspec (Spec, describe, it, shouldBe)
+import Types (GlobFilter (GlobFilter), LicenseScanPathFilters (..))
 
 recursiveArchive :: FixtureArtifact
 recursiveArchive =
@@ -65,7 +66,7 @@ spec = do
     it "should find licenses in nested archives" $ do
       extractedDir <- getArtifact recursiveArchive
       let scanDir = extractedDir </> [reldir|cli-license-scan-integration-test-fixtures-main/recursive-archive|]
-      units <- runStack . runDiagnostics . ignoreStickyLogger . runExecIO . runReadFSIO . fmap licenseSourceUnitLicenseUnits $ scanVendoredDep scanDir vendoredDep
+      units <- runStack . runDiagnostics . ignoreStickyLogger . runExecIO . runReadFSIO . fmap licenseSourceUnitLicenseUnits $ scanVendoredDep scanDir Nothing vendoredDep
       PIO.removeDirRecur extractedDir
       case units of
         Failure ws eg -> fail (show (renderFailure ws eg "An issue occurred"))
@@ -77,5 +78,21 @@ spec = do
           where
             mitUnit :: LicenseUnit
             mitUnit = fromMaybe emptyLicenseUnit (head' $ NE.filter (\u -> licenseUnitName u == "mit") us)
+            apacheUnit :: LicenseUnit
+            apacheUnit = fromMaybe emptyLicenseUnit (head' $ NE.filter (\u -> licenseUnitName u == "apache-2.0") us)
+
+    it "should filter licenses if LicenseScanPathFilters is set" $ do
+      extractedDir <- getArtifact recursiveArchive
+      let scanDir = extractedDir </> [reldir|cli-license-scan-integration-test-fixtures-main/recursive-archive|]
+      let licenseScanPathFilters = LicenseScanPathFilters{licenseScanPathFiltersOnly = [GlobFilter "**.rb"], licenseScanPathFiltersExclude = []}
+      units <- runStack . runDiagnostics . ignoreStickyLogger . runExecIO . runReadFSIO . fmap licenseSourceUnitLicenseUnits $ scanVendoredDep scanDir (Just licenseScanPathFilters) vendoredDep
+      PIO.removeDirRecur extractedDir
+      case units of
+        Failure ws eg -> fail (show (renderFailure ws eg "An issue occurred"))
+        Success _ us -> do
+          length us `shouldBe` 1
+          NE.sort (NE.map licenseUnitName us) `shouldBe` NE.fromList ["apache-2.0"]
+          NE.sort (licenseUnitFiles apacheUnit) `shouldBe` NE.fromList ["vendor/foo/bar/bar_apache.rb", "vendor/foo/bar/baz/something.rb"]
+          where
             apacheUnit :: LicenseUnit
             apacheUnit = fromMaybe emptyLicenseUnit (head' $ NE.filter (\u -> licenseUnitName u == "apache-2.0") us)