+++ src/App/Fossa/Analyze.hs
@@ -268,9 +268,11 @@ analyze cfg = Diag.context "fossa-analyze" $ do
             analyzeVSI basedir revision filters skipResolutionSet
         _ -> pure Nothing
   dynamicLinkedResults <-
-    Diag.errorBoundaryIO . diagToDebug $
-      Diag.context "discover-dynamic-linking" . doAnalyzeDynamicLinkedBinary basedir . Config.dynamicLinkingTarget $
-        Config.vsiOptions cfg
+    Diag.errorBoundaryIO
+      . diagToDebug
+      . runReader filters
+      $ Diag.context "discover-dynamic-linking" . doAnalyzeDynamicLinkedBinary basedir . Config.dynamicLinkingTarget
+      $ Config.vsiOptions cfg
   binarySearchResults <-
     Diag.errorBoundaryIO . diagToDebug $
       Diag.context "discover-binaries" $
@@ -406,6 +408,7 @@ doAnalyzeDynamicLinkedBinary ::
   , Has Logger sig m
   , Has ReadFS sig m
   , Has Exec sig m
+  , Has (Reader AllFilters) sig m
   ) =>
   Path Abs Dir ->
   DynamicLinkInspect ->