+++ test/Fossa/API/TypesSpec.hs
@@ -2,7 +2,7 @@ module Fossa.API.TypesSpec (spec) where
 
 import Data.Aeson (FromJSON, ToJSON, fromJSON, toJSON)
 import Data.Text (Text)
-import Fossa.API.Types (Issue (..), IssueRule (..), IssueType (..), Issues (..))
+import Fossa.API.Types (Issue (..), IssueRule (..), IssueSummaryRevision (..), IssueSummaryTarget (..), IssueType (..), Issues (..), IssuesSummary (..))
 import Hedgehog.Gen qualified as Gen
 import Hedgehog.Range qualified as Range
 import Test.Hspec
@@ -23,6 +23,26 @@ genIssues =
     <$> Gen.int (Range.linear minBound maxBound)
     <*> Gen.list (Range.linear 0 100) genIssue
     <*> arbitraryText
+    <*> Gen.maybe genIssueRevisionSummary
+
+genIssueRevisionSummary :: Gen IssuesSummary
+genIssueRevisionSummary =
+  IssuesSummary
+    <$> genIssueRevision
+    <*> Gen.maybe (Gen.list (Range.linear 0 100) genIssueRevisionTargets)
+
+genIssueRevision :: Gen IssueSummaryRevision
+genIssueRevision =
+  IssueSummaryRevision
+    <$> arbitraryText
+    <*> arbitraryText
+    <*> Gen.maybe (Gen.bool)
+
+genIssueRevisionTargets :: Gen IssueSummaryTarget
+genIssueRevisionTargets =
+  IssueSummaryTarget
+    <$> arbitraryText
+    <*> Gen.list (Range.linear 0 5) arbitraryText
 
 genIssue :: Gen Issue
 genIssue =
@@ -33,6 +53,10 @@ genIssue =
     <*> arbitraryText
     <*> genIssueType
     <*> Gen.maybe genIssueRule
+    <*> Gen.maybe arbitraryText
+    <*> Gen.maybe arbitraryText
+    <*> Gen.maybe arbitraryText
+    <*> Gen.maybe arbitraryText
 
 genIssueType :: Gen IssueType
 genIssueType =
@@ -46,7 +70,8 @@ genIssueType =
     ]
 
 genIssueRule :: Gen IssueRule
-genIssueRule = IssueRule <$> Gen.maybe arbitraryText
+genIssueRule =
+  IssueRule <$> Gen.maybe (Gen.int (Range.linear 0 100))
 
 arbitraryText :: Gen Text
 arbitraryText = Gen.text (Range.linear 0 100) Gen.unicodeAll