+++ vendor_download.sh
@@ -81,7 +81,7 @@ rm $WIGGINS_RELEASE_JSON
 echo "Wiggins download successful"
 echo
 
-THEMIS_TAG="2022-10-11-40ed95c"
+THEMIS_TAG="2023-01-13-330c1a9-1673649513"
 echo "Downloading themis binary"
 echo "Using themis release: $THEMIS_TAG"
 THEMIS_RELEASE_JSON=vendor-bins/themis-release.json